code prettier

+++ cfiles/convertions.c
@@ -1,12 +1,12 @@
 
 #include "../cub3d.h"
 
-double	deg2rad(double degrees)
+float	deg2rad(double degrees)
 {
 	return (degrees * (PI / 180.0));
 }
 
-double	rad2deg(double radians)
+float	rad2deg(double radians)
 {
 	return (radians * (180.0 / PI));
 }
